web /backend_flask/requirements.txt
aiodns==3.1.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.1.0
backtrader==1.9.78.123
beautifulsoup4==4.12.3
blinker==1.8.2
catboost==1.2.5
ccxt==4.1.75
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
dateparser==1.2.0
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.46.0
frozenlist==1.4.0
google==3.0.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lightgbm==4.4.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.6.0
numpy==1.26.2
openai==1.36.1
openpyxl==3.1.5
orjson==3.10.6
packaging==23.2
pandas==2.1.3
pandas-ta==0.3.14b0
parso==0.8.4
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.2.2
plotly==5.22.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.4.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.0.3
regex==2023.10.3
requests==2.31.0
rich==13.7.1
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.13.2
setuptools==69.0.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
TA-Lib==0.4.28
tenacity==8.5.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.8.0
urllib3==2.1.0
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
xgboost==2.1.0
yarl==1.9.3
yellowbrick==1.5
